renamed packages to model, controller,view. also moved WaitingListRepository into model.
