+++ app/src/main/java/com/example/releasethekraken/controller/WaitingListService.java
@@ -1,0 +1,4 @@
+package com.example.releasethekraken.controller;
+
+public class WaitingListService {
+}
