--- app/src/main/java/com/example/releasethekraken/data/WaitingListRepository.java
@@ -1,4 +1,0 @@
-package com.example.releasethekraken.data;
-
-public class WaitingListRepository {
-}
