--- app/src/main/java/com/example/releasethekraken/logic/WaitingListService.java
@@ -1,4 +1,0 @@
-package com.example.releasethekraken.logic;
-
-public class WaitingListService {
-}
