+++ app/src/main/java/com/example/releasethekraken/model/WaitingListRepository.java
@@ -1,0 +1,4 @@
+package com.example.releasethekraken.model;
+
+public class WaitingListRepository {
+}
