--- app/src/main/java/com/example/releasethekraken/models/WaitingListEntry.java
@@ -1,4 +1,0 @@
-package com.example.releasethekraken.models;
-
-public class WaitingListEntry {
-}
